수정

+++ src/main/java/com/sunny/backend/friends/service/FriendService.java
@@ -4,6 +4,7 @@ import static com.sunny.backend.friends.exception.FriendErrorCode.*;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 
@@ -35,6 +36,7 @@ public class FriendService {
 		return FriendListResponse.of(friends);
 	}
 
+	@Transactional
 	public void addFriend(CustomUserPrincipal customUserPrincipal, Long userFriendId) throws IOException {
 		Users user = customUserPrincipal.getUsers();
 		user.canNotMySelf(userFriendId);
@@ -61,7 +63,7 @@ public class FriendService {
 		String title = "[SUNNY] " + sendFriend.getUsers().getNickname();
 		String body = "님이 친구를 신청했어요!";
 		String bodyTitle = "친구 신청을 받았어요";
-		friendNotiService.sendNotifications(title, body, bodyTitle, sendFriend);
+		friendNotiService.sendNotifications(title, body, bodyTitle, receiveFriend);
 	}
 
 	@Transactional
@@ -74,19 +76,20 @@ public class FriendService {
 		// 상대 친구 관계를 조회, 승인하는 경우
 		// 친구 관계가 존재하는 경우 SEND -> FRIEND
 		// 친구 관계가 존재하지 않는 경우는 서버 문제임으로 받은 친구 관계 삭제 후 에러 처리
-		friendRepository.findByUsersAndUserFriend(receiveFriend.getUserFriend(), receiveFriend.getUsers())
-			.ifPresentOrElse(
-				sendFriend -> sendFriend.updateFriendStatus(FriendStatus.FRIEND),
-				() -> {
-					friendRepository.delete(receiveFriend);
-					throw new CustomException(FRIEND_SERVER_ERROR);
-				}
-			);
+		Optional<Friend> friendOptional = friendRepository.findByUsersAndUserFriend(receiveFriend.getUserFriend(),
+			receiveFriend.getUsers());
+		if (friendOptional.isEmpty()) {
+			friendRepository.delete(receiveFriend);
+			throw new CustomException(FRIEND_SERVER_ERROR);
+		}
+
+		Friend sendFriend = friendOptional.get();
+		sendFriend.updateFriendStatus(FriendStatus.FRIEND);
 
 		String title = "[SUNNY] " + receiveFriend.getUsers().getNickname();
 		String body = "님이 친구 신청을 수락했어요";
 		String bodyTitle = "친구 신청 결과를 알려드려요";
-		friendNotiService.sendNotifications(title, body, bodyTitle, receiveFriend);
+		friendNotiService.sendNotifications(title, body, bodyTitle, sendFriend);
 		receiveFriend.updateFriendStatus(FriendStatus.FRIEND);
 	}
 
@@ -100,19 +103,20 @@ public class FriendService {
 		// 상대 친구 관계를 조회, 거절하는 경우
 		// 친구 관계가 존재하는 경우 상대 친구 관계도 같이 삭제
 		// 친구 관계가 존재하지 않는 경우는 서버 문제임으로 받은 친구 관계 삭제 후 에러 처리
-		friendRepository.findByUsersAndUserFriend(receiveFriend.getUserFriend(), receiveFriend.getUsers())
-			.ifPresentOrElse(
-				friendRepository::delete,
-				() -> {
-					friendRepository.delete(receiveFriend);
-					throw new CustomException(FRIEND_SERVER_ERROR);
-				}
-			);
+		Optional<Friend> friendOptional = friendRepository.findByUsersAndUserFriend(receiveFriend.getUserFriend(),
+			receiveFriend.getUsers());
+		if (friendOptional.isEmpty()) {
+			friendRepository.delete(receiveFriend);
+			throw new CustomException(FRIEND_SERVER_ERROR);
+		}
+		Friend sendFriend = friendOptional.get();
+
 		String title = "[SUNNY] " + receiveFriend.getUsers().getNickname();
 		String body = "님이 친구 신청을 거절했어요";
 		String bodyTitle = "친구 신청 결과를 알려드려요";
-		friendNotiService.sendNotifications(title, body, bodyTitle, receiveFriend);
+		friendNotiService.sendNotifications(title, body, bodyTitle, sendFriend);
 
+		friendRepository.delete(friendOptional.get());
 		friendRepository.delete(receiveFriend);
 	}
